This is great stuff here.
Great stuff man

+++ Library/src/controller/Controller.java
@@ -1,15 +1,20 @@
 package controller;
 
 import java.sql.SQLException;
+import java.sql.SQLIntegrityConstraintViolationException;
+import java.util.Optional;
 
 import com.mysql.cj.x.protobuf.MysqlxDatatypes.Scalar.String;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import model.Book;
 
@@ -53,16 +58,37 @@ public class Controller {
 
 	@FXML
 	void handleDelete(ActionEvent event) {
-		try {
-			myBook.setRowID(tblMain.getSelectionModel().getSelectedItem().getRowID());
-			myBook.delete();
-			tableRfsh();
-			handleClear();
-		} catch (Exception e) {
-			e.printStackTrace();
+		if (tblMain.getSelectionModel().isEmpty()) {
+			Alert empty = new Alert(AlertType.ERROR);
+			empty.setTitle("Error");
+			empty.setHeaderText("No Row Selected");
+			empty.setContentText("YOU HAVE TO SELECT SOMETHING TO DELETE IT YOU IMBECILE!!!!");
+			empty.showAndWait();
+		} else {
+			try {
+				Alert myAlert = new Alert(AlertType.CONFIRMATION);
+				myAlert.setTitle("Confirm Delete");
+				myAlert.setHeaderText("Are you sure?");
+				myAlert.setContentText("Are your sure you wish to do this?");
+				Optional<ButtonType> answer = myAlert.showAndWait();
+				if (answer.isPresent() && answer.get().equals(ButtonType.OK)) {
+					myBook.setRowID(tblMain.getSelectionModel().getSelectedItem().getRowID());
+					myBook.delete();
+					tableRfsh();
+					handleClear();
+				} else {
+					Alert cancelled = new Alert(AlertType.INFORMATION);
+					cancelled.setTitle("Delete Cancelled");
+					cancelled.setHeaderText("Fool");
+					cancelled.setContentText("COWARD!");
+					cancelled.showAndWait();
+				}
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
 		}
 	}
-	
+
 	@FXML
 	void handleUpdate(ActionEvent event) {
 		try {
@@ -74,6 +100,12 @@ public class Controller {
 			myBook.update();
 			tableRfsh();
 			handleClear();
+		} catch (NullPointerException e) {
+			Alert empty = new Alert(AlertType.ERROR);
+			empty.setTitle("Error");
+			empty.setHeaderText("No Row Selected");
+			empty.setContentText("Update what?");
+			empty.showAndWait();
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
@@ -89,9 +121,33 @@ public class Controller {
 			myBook.create();
 			tableRfsh();
 			handleClear();
+		} catch (SQLIntegrityConstraintViolationException e) {
+			Alert empty = new Alert(AlertType.ERROR);
+			empty.setTitle("Error");
+			empty.setHeaderText("Nothing To Save");
+			empty.setContentText("Are you braindead? Enter something to save.");
+			empty.showAndWait();
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
+	}
+
+	@FXML
+	void handleSearch(ActionEvent event) {
+		Alert empty = new Alert(AlertType.INFORMATION);
+		empty.setTitle("No");
+		empty.setHeaderText("Nope.");
+		empty.setContentText("no...");
+		empty.showAndWait();
+	}
+
+	@FXML
+	void error() {
+		Alert empty = new Alert(AlertType.ERROR);
+		empty.setTitle("Error");
+		empty.setHeaderText("Error");
+		empty.setContentText("Error");
+		empty.showAndWait();
 	}
 
 	@FXML
